go: move to s3secrets-helper directory; resulting binary name changed.
