--- .gitignore
@@ -1,2 +1,0 @@
-# binary
-/elastic-ci-stack-s3-secrets-hooks
