+++ README.md
@@ -12,12 +12,12 @@ Different types of secrets are supported and exposed to your builds in appropria
 
 The hooks needs to be installed directly in the agent so that secrets can be downloaded before jobs attempt checking out your repository. We are going to assume that buildkite has been installed at `/buildkite`, but this will vary depending on your operating system. Change the instructions accordingly.
 
-The core of the hook is a Go binary, the source of which is in this repository. `go build` yields a `git-credential-s3-secrets` binary which must be placed in `$PATH`. Or, look for a precompiled binary a https://github.com/buildkite/elastic-ci-stack-s3-secrets-hooks/releases (coming soon?).
+The core of the hook is an `s3secrets-helper` binary, the result of `go build` in the `s3secrets-helper/` subdirectory of this repository. It must be placed in `$PATH` to be found by the `hooks/environment` wrapper script.
 
 ```bash
 # clone to a path your buildkite-agent can access
 git clone https://github.com/buildkite-plugins/s3-secrets-buildkite-plugin.git /buildkite/s3_secrets
-(cd /buildkite/s3_secrets && go install)
+(cd /buildkite/s3_secrets/s3secrets-helper && go build -o /usr/local/bin/s3secrets-helper)
 ```
 
 Modify your agent's global hooks (see [https://buildkite.com/docs/agent/v3/hooks#global-hooks](https://buildkite.com/docs/agent/v3/hooks#global-hooks)):
