--- go.mod
@@ -1,5 +1,0 @@
-module github.com/buildkite/elastic-ci-stack-s3-secrets-hooks
-
-go 1.15
-
-require github.com/aws/aws-sdk-go v1.35.14
