+++ s3secrets-helper/.gitignore
@@ -1,0 +1,2 @@
+# default `go build` binary
+/s3secrets-helper
