+++ s3secrets-helper/go.mod
@@ -1,0 +1,5 @@
+module github.com/buildkite/elastic-ci-stack-s3-secrets-hooks/s3secrets-helper/v2
+
+go 1.15
+
+require github.com/aws/aws-sdk-go v1.35.14
